python-tutorials/Airflow/pyproject.toml
[tool.poetry]
name = "airflow-env"
version = "0.1.0"
description = ""
authors = ["stephenjera"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
apache-airflow = "2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
